new hints component, supports rows
