--- components/nonogram/row-hints.tsx
@@ -1,9 +1,0 @@
-import {Tables} from "@/types/database.types";
-import {NonogramGrid} from "@/types/types";
-
-export function RowHints({nonogram, grid}: {
-    nonogram: Tables<"nonograms">;
-    grid: NonogramGrid;
-}) {
-    return null;
-}
